Merge upstream changes to Pix4D fork

+++ bin/dry-run.rb
@@ -63,6 +63,7 @@ require "json"
 require "byebug"
 require "logger"
 require "dependabot/logger"
+require "stackprof"
 
 Dependabot.logger = Logger.new($stdout)
 
@@ -201,6 +202,11 @@ option_parse = OptionParser.new do |opts|
           "Only update vulnerable dependencies") do |_value|
     $options[:security_updates_only] = true
   end
+
+  opts.on("--profile",
+          "Profile using Stackprof. Output in `tmp/stackprof-<datetime>.dump`") do
+    $options[:profile] = true
+  end
 end
 
 option_parse.parse!
@@ -272,9 +278,7 @@ def cached_dependency_files_read
   )
   FileUtils.mkdir_p(cache_dir) unless Dir.exist?(cache_dir)
 
-  if File.exist?(cache_manifest_path)
-    cached_manifest = File.read(cache_manifest_path)
-  end
+  cached_manifest = File.read(cache_manifest_path) if File.exist?(cache_manifest_path)
   cached_dependency_files = JSON.parse(cached_manifest) if cached_manifest
 
   all_files_cached = cached_dependency_files&.all? do |file|
@@ -321,9 +325,7 @@ def cached_dependency_files_read
     end
     # Initialize a git repo so that changed files can be diffed
     if $options[:write]
-      if File.exist?(".gitignore")
-        FileUtils.cp(".gitignore", File.join(cache_dir, ".gitignore"))
-      end
+      FileUtils.cp(".gitignore", File.join(cache_dir, ".gitignore")) if File.exist?(".gitignore")
       Dir.chdir(cache_dir) do
         system("git init . && git add . && git commit --allow-empty -m 'Init'")
       end
@@ -425,6 +427,8 @@ def handle_dependabot_error(error:, dependency:)
   puts " => handled error whilst updating #{dependency.name}: #{error_details.fetch(:'error-type')} "\
        "#{error_details.fetch(:'error-detail')}"
 end
+
+StackProf.start(raw: true) if $options[:profile]
 
 source = Dependabot::Source.new(
   provider: $options[:provider],
@@ -690,6 +694,10 @@ dependencies.each do |dep|
 rescue StandardError => e
   handle_dependabot_error(error: e, dependency: dep)
 end
+
+StackProf.stop if $options[:profile]
+StackProf.results("tmp/stackprof-#{Time.now.strftime('%Y-%m-%d-%H:%M')}.dump") if $options[:profile]
+
 # rubocop:enable Metrics/BlockLength
 
 # rubocop:enable Style/GlobalVars
